+++ controllers/middleware/modules.js
@@ -41,7 +41,7 @@ exports._md.usePassport = [passport.initialize(), passport.session()];
 // authentication
 exports._md.isSignedIn = [exports._md.doNotCrawled, ...exports._md.usePassport, (req, res, next) => {
     if (req.isAuthenticated()) return next();
-    if (res.locals._view.flash.pass[0]) {
+    if (res.locals._view.flash.pass[0] === undefined) {
         req.flash('error', String(res.locals._view.flash.error));
         req.flash('info', String(res.locals._view.flash.info));
     } else req.flash('error', 'Please sign in first!');
@@ -52,9 +52,9 @@ exports._md.isSignedIn = [exports._md.doNotCrawled, ...exports._md.usePassport, 
 
 // authorization
 exports._md.isAuthorized = [...exports._md.isSignedIn, async (req, res, next) => {
-    const [field, val] = req.params.canonical
+    const [field, val] = req.params.canonical !== undefined
         ? ['canonical', req.params.canonical]
-        : ['_id', _fn.string.readMongoID(req.url)];
+        : ['_id', _fn.string.readMongoId(req.url)];
     const count = await require('../../models').postModel.count({ [field]: val, 'author._id': req.user });
     if (count !== 1) {
         req.flash('error', 'You do not have a valid authorization...');